[REMANIEMENT] Déplace le composant Modale dans son propre fichier
