+++ mon-aide-cyber-ui/src/composants/modale/PortailModale.tsx
@@ -1,83 +1,22 @@
 import {
   ContexteModale,
   ElementModale,
-  TailleModale,
 } from '../../fournisseurs/ContexteModale.ts';
 import { createPortal } from 'react-dom';
 import {
-  ForwardedRef,
-  forwardRef,
   PropsWithChildren,
-  ReactElement,
   useCallback,
   useEffect,
   useRef,
   useState,
 } from 'react';
-
-type ProprietesElementModale = ElementModale & {
-  boutonFermer: ReactElement;
-  surClickEnDehors: () => void;
-};
-
-const taillesModale: Map<TailleModale, string> = new Map<TailleModale, string>([
-  ['centree', 'fr-col-xl-6 fr-col-8'],
-  ['moyenne', 'fr-col-xl-7 fr-col-10'],
-  ['large', 'fr-col-12'],
-]);
-
-export const Modale = forwardRef(function Modale(
-  proprietes: PropsWithChildren<ProprietesElementModale>,
-  referenceDialogue: ForwardedRef<any>
-) {
-  const ref = useRef<HTMLDivElement | null>(null);
-
-  const clickEnDehors = (e: MouseEvent) => {
-    if (ref.current && !ref.current.contains(e.target as Node)) {
-      proprietes.surClickEnDehors();
-    }
-  };
-
-  useEffect(() => {
-    addEventListener('click', clickEnDehors, true);
-
-    return () => {
-      removeEventListener('click', clickEnDehors);
-    };
-  });
-
-  const taille =
-    taillesModale.get(proprietes.taille || 'centree') || 'fr-col-6';
-
-  return (
-    <div className="fr-container fr-container--fluid">
-      <div className="fr-grid-row fr-grid-row--center">
-        <div ref={ref} className={taille}>
-          <div
-            ref={referenceDialogue}
-            className="fr-modal__body modale-mac fr-m-0 fr-p-0"
-          >
-            <div className="fr-modal__header">
-              <div className="fr-grid-row fr-col-12">
-                <div className="fr-col-10">
-                  <h4 id="titre-modale">{proprietes.titre}</h4>
-                </div>
-                <div className="fr-col-2">{proprietes.boutonFermer}</div>
-              </div>
-            </div>
-            <div className="fr-modal__content">{proprietes.corps}</div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-});
+import { Modale } from './Modale.tsx';
 
 export const PortailModale = ({ children }: PropsWithChildren) => {
   const [elementModale, setElementModale] = useState<null | ElementModale>(
     null
   );
-  const [classModale, setClasseModale] = useState<string | undefined>(
+  const [classeModale, setClasseModale] = useState<string | undefined>(
     undefined
   );
   const [ariaModale, setAriaModale] = useState<boolean>(false);
@@ -155,11 +94,10 @@ export const PortailModale = ({ children }: PropsWithChildren) => {
       <dialog
         aria-labelledby="titre-modale"
         id="modale"
-        className={`fr-modal ${classModale}`}
+        className={`fr-modal ${classeModale}`}
         aria-modal={ariaModale}
         open={modaleOuverte}
-        role="dialog"
-      ></dialog>
+      />
       {elementModale
         ? createPortal(
             <Modale
